create function to discretize pixels to discrete list
