+++ models/utils/get_n_worst_predictions.py
@@ -4,7 +4,7 @@ from sklearn.model_selection import train_test_split
 from tensorflow.keras import layers
 
 from models.transformer import create_vit_model, fit_vit_model
-from models.utils.get_test_data import get_average_green_band_data
+from models.utils.get_train_data import get_average_green_band_data
 from models.utils.root_mean_squared_error import root_mean_squared_error
 import os.path as osp
 import csv
